pyproject.toml
[project]
name = "roaringcal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.0",
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
